+++ examples/alanine/ala_dipeptide_pathmovertest.py
@@ -10,18 +10,20 @@ TIS simulation on alanine dipeptide.
 # script
 
 import logging.config
+import time
+import sys
+import os
+
 import numpy as np
 import mdtraj as md
-import time
 
-import sys, os
 sys.path.append(os.path.abspath('../'))
 sys.path.append(os.path.abspath('../../'))
 
 # in principle, all of these imports should be simplified once this is a
 # package
-from openpathsampling.collectivevariable import CV_Function, CV_Volume, CV_MD_Function
-from openpathsampling.openmm_engine import OpenMMEngine
+from openpathsampling.collectivevariable import CV_Volume, CV_MD_Function
+from openpathsampling.dynamics.openmm.openmm_engine import OpenMMEngine
 from openpathsampling.snapshot import Snapshot
 from openpathsampling.volume import LambdaVolumePeriodic, VolumeFactory as vf
 from openpathsampling.pathmover import PathMoverFactory as mf
