+++ examples/alanine/ala_dipeptide_simple.py
@@ -9,25 +9,26 @@ TIS simulation on alanine dipeptide.
 # that more of what the user will need to do will be in the __main__ of this
 # script
 
+import time
+import sys
+import os
+
 import numpy as np
 import mdtraj as md
-import time
 
-import sys, os
 sys.path.append(os.path.abspath('../'))
 sys.path.append(os.path.abspath('../../'))
 
 # in principle, all of these imports should be simplified once this is a
 # package
-from openpathsampling.collectivevariable import CV_Function, CV_Volume
-from openpathsampling.openmm_engine import OpenMMEngine
+from openpathsampling.collectivevariable import CV_Function
+from openpathsampling.dynamics.openmm.openmm_engine import OpenMMEngine
 from openpathsampling.snapshot import Snapshot
 from openpathsampling.volume import LambdaVolumePeriodic, VolumeFactory as vf
 from openpathsampling.pathmover import PathMoverFactory as mf
 from openpathsampling.ensemble import EnsembleFactory as ef
 from openpathsampling.ensemble import (LengthEnsemble, SequentialEnsemble, AllOutEnsemble,
                               AllInEnsemble)
-from openpathsampling.pathsimulator import Bootstrapping
 from openpathsampling.pathmover import PathMover
 from openpathsampling.shooting import UniformSelector
