+++ openpathsampling/tests/test_toy_dynamics.py
@@ -3,17 +3,16 @@
 '''
 
 import os
-import numpy as np
 
 from nose.tools import (assert_equal, assert_not_equal, assert_items_equal,
                         assert_almost_equal, raises)
-from nose.plugins.skip import Skip, SkipTest
-from test_helpers import true_func, assert_equal_array_array
-
+from nose.plugins.skip import SkipTest
 from openpathsampling.toy_dynamics.toy_pes import *
-from openpathsampling.toy_dynamics.toy_integrators import *
 from openpathsampling.toy_dynamics.toy_engine import *
-from openpathsampling.topology import ToyTopology
+
+from test_helpers import true_func, assert_equal_array_array
+from openpathsampling.dynamics.toy_dynamics.toy_integrators import *
+from openpathsampling.dynamics.toy.topology import ToyTopology
 from openpathsampling.snapshot import Snapshot, Momentum, Configuration
