+++ openpathsampling/tests/testopenmmengine.py
@@ -1,20 +1,19 @@
 '''
 @author David W.H. Swenson
 '''
-import os
-from nose.tools import (assert_equal, assert_not_equal, assert_items_equal,
-                        assert_almost_equal, raises)
-from nose.plugins.skip import Skip, SkipTest
+import time
+
+from nose.tools import (assert_equal, assert_items_equal)
+from nose.plugins.skip import SkipTest
+import simtk.unit as u
+
 from test_helpers import (true_func, data_filename,
                           assert_equal_array_array,
                           assert_not_equal_array_array)
-
-from openpathsampling.openmm_engine import *
+from openpathsampling.dynamics.openmm.openmm_engine import *
 from openpathsampling.snapshot import Snapshot
 from openpathsampling.snapshot import Momentum, Configuration
 
-import simtk.unit as u
-import time
 
 class testOpenMMEngine(object):
     def setUp(self):
